changed to size_t and added parameter set convenience struct

+++ src/layer.h
@@ -17,7 +17,7 @@ typedef enum LAYER_TYPE_ {
 // strictly as input
 typedef struct Layer_ {
     LAYER_TYPE type;
-    int size;
+    size_t size;
     void (*activation)(Matrix*);
     Matrix* input; // (num_examples x size)
 } Layer;
@@ -30,7 +30,7 @@ typedef struct Connection_ {
 } Connection;
 
 // returns layer given metadata and configuration
-Layer* createLayer(LAYER_TYPE type, int size, void (*activation)(Matrix*));
+Layer* createLayer(LAYER_TYPE type, size_t size, void (*activation)(Matrix*));
 
 // creates a connection and creates weights and bias matrices
 Connection* createConnection(Layer* from, Layer* to);
@@ -51,7 +51,7 @@ void destroyConnection(Connection* connection);
 /*
     Begin functions.
 */
-Layer* createLayer(LAYER_TYPE type, int size, void (*activation)(Matrix*)){
+Layer* createLayer(LAYER_TYPE type, size_t size, void (*activation)(Matrix*)){
     Layer* layer = (Layer*)malloc(sizeof(Layer));
     layer->type = type;
     layer->size = size;
